requirements.txt
streamlit
qrcode
Pillow
requests